[feaLib] Don't try to combine ligature & multisub rules
I'm not sure that this patch is the correct solution to this problem,
but it is _a_ solution.

Specifically the problem is that the AnySubstBuilder can generate
multiple lookups, but when this happens those extra lookups do not get
added to the appropriate feature. My solution here is to force a new
lookup whenever we see that the AnySubstBuilder would otherwise contain
mixed multiple/ligature sub rules, which cannot be combined anyway.

Open to alternatives, but this does fix the issue.