[cu2qu] Handle degenerate case of 3 equal points
Dealing with a situation where open corner erasure is producing a cubic
segment where all the points are equal, which is subsequently falling
through the conversion to quadratics.

This handles that situation naively, with a branch; actually improving
the robustness of the numerical code is an exercise for someone else.