Fix interpolation, and allow for common operations on pre-compiled fonts
(aims to facilitate googlefonts/ufo2ft#945)

Previously, calculating an individual interpolated value did not account
for avar mapping. This commit fixes this, as well as allowing for
interpolation and other common operations to be performed on fonts prior
to compilation by decoupling from the exact ttLib representations of
axes.

Some refactoring has been undergone to assist with this, including
adding some more types and docs, and splitting the builder into a
separate class to limit the amount of state in the VariableScalar class
itself.

The fonttools and ufo2ft test suites pass with these changes, although
there is no coverage for the designspace-based interpolation yet. This
should be introduced before merging.