Don’t use deprecated uharfbuzz API
Depend on a recent enough uharfbuzz to be able to unconditionally use
serialize_with_tag(), and drop the code that tries to use repack() if
repack_with_tag() is missing.

Using ancient versions of uharfbuzz is not a good idea anyway, as this
misses later fixes to the repacking code.

Fixes https://github.com/fonttools/fonttools/issues/3917